adds Assertions.maps_equal/3 for simple true/false comparison
This will fail:

assert_list_equal(list1, list2, &assert_maps_equal(&1, &2, keys))

when the elements of list1 are not in the same order as list2.  This is
because assert_maps_equal/3 raises an error when two elements are not
equal.

assert_list_equal(list1, list2, &maps_equal(&1, &2, keys))

will let assert_list_equal/3 work the way it's supposed to.